requirements.txt

apache-beam[gcp]
google-cloud-pubsub
google-cloud-bigquery
